+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Add Diagram.Note helper method
 - Label shape supports linebreaks
 - Add shapes Process, Store